% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm_matching.R
\name{is.qm_matching}
\alias{is.qm_matching}
\title{Check \code{qm_matching} object}
\usage{
is.qm_matching(obj)
}
\arguments{
\item{obj}{object to check.}
}
\value{
Returns \code{TRUE} if \code{obj} is valid,
        otherwise \code{FALSE}.
}
\description{
\code{is.qm_matching} checks whether the provided object
is a valid instance of the \code{\link{qm_matching}} class.
It does not check whether the matching it describes is sensible.
}

