% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{match}
\alias{match}
\title{Construct matchings}
\usage{
match(distance_object, treatments, treatment_constraints = NULL,
  total_size_constraint = NULL, caliper = NULL, subset = NULL, ...)
}
\arguments{
\item{distance_object}{abc}

\item{treatments}{abc}

\item{treatment_constraints}{abc}

\item{total_size_constraint}{abc}

\item{caliper}{abc}

\item{subset}{abc}

\item{...}{abc}
}
\value{
abc
}
\description{
abc
}

