% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{potential_outcomes}
\alias{potential_outcomes}
\title{Average potential outcome estimator in matched groups}
\usage{
potential_outcomes(outcomes, treatments, matching, targets = NULL,
  subset = NULL)
}
\arguments{
\item{outcomes}{numeric vector with observed outcomes.}

\item{treatments}{integer or factor vector with treatment indicators.}

\item{matching}{\code{\link{qm_matching}} or \code{\link[Rscclust]{Rscc_clustering}}
object describing the matched groups.}

\item{targets}{vector of treatment labels (corresponding to \code{treatments})
specifying which potential outcomes to estimate. If \code{NULL},
all potential outcomes will be estimated.}

\item{subset}{units to estimate potential outcomes for. If \code{NULL},
the estimate will pertain to all units in the sample (i.e.,
corresponding to ATE). A non-null value specificies a subset
of units that the estimate should pertain to (e.g., ATT or ATC).
If \code{subset} is a logical vector and of the same length as the
sample size, units indicated with \code{TRUE} will be included.
Otherwise, \code{subset} should contain treatment labels, and
the corresponding units (as given by \code{treatments}) will be
included.}
}
\value{
Returns a named numeric vector with estimates of the potential outcomes.
}
\description{
\code{potential_outcomes} estimates potential outcomes in matched groups. Provided
matched groups, outcomes and treatment indicators, the function returns estimates of the average
potential outcomes for all treatments for the units in the sample. It is also possible
to estimate potential outcomes for subsets of the units. For example, one can estimate
the potential outcomes for units assigned to a certain treatment condition.
}
\examples{
# Example input
outcomes <- (1:100)^0.5
treatments <- factor(rep(c("a", "b", "c", "c"), each = 25))
matching <- qm_matching(rep(1:10, 10))

# Estimate all potential outcomes for the whole sample
potential_outcomes(outcomes, treatments, matching)

# Estimate only the average potential outcome for the "a" treatment
potential_outcomes(outcomes, treatments, matching, targets = "a")

# Estimate potential outcomes for units assigned to treatments "b" or "c"
potential_outcomes(outcomes, treatments, matching, subset = c("b", "c"))

# Estimate potential outcomes for units 1:15 and 31:45
subset <- c(rep(TRUE, 15), rep(FALSE, 15), rep(TRUE, 15), rep(FALSE, 55))
potential_outcomes(outcomes, treatments, matching, subset = subset)

}
\seealso{
See \code{\link{treatment_effects}} for a treatment effects estimator.
}

