% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickmatch-package.R
\docType{package}
\name{quickmatch-package}
\alias{quickmatch-package}
\title{quickmatch: Fast Matching in Large Data Sets with Complicated Constraints}
\description{
\code{quickmatch} provides matching functions for constructing generalized full matchings
in large samples. It also provides computationally efficient estimators for
treatment effects and potential outcomes in matched groups. See
\code{\link{quickmatch}} for the core matching function. See
\code{\link{treatment_effects}} and \code{\link{potential_outcomes}}
for estimators.
}
\details{
The package is still under development. Please use it with caution.

More information and the latest version is found here:
\url{https://github.com/fsavje/quickmatch}.

Bug reports and suggestions are greatly appreciated. They
are best reported here:
\url{https://github.com/fsavje/quickmatch/issues}.
}

