% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickmatch7.R
\docType{methods}
\name{quickmatch}
\alias{quickmatch}
\title{Quickmatch algorithm}
\usage{
quickmatch(x, z, caliper, controls = 1L, data, within, method = "nno")
}
\arguments{
\item{x}{Numeric vector with the scores of treated and control objects,
or formula, or glm.}

\item{z}{If x is a numeric vector, z is a vector with 1 for treated objects}

\item{caliper}{The caliper, i.e., the maximal distance allowed between}

\item{controls}{The maximal number of controls matched to a single treated object}

\item{data}{An optional data frame, list or environment for arguments x,z
and within.}

\item{within}{Vector of factors dividing objects into groups.}

\item{method}{Method for matching: "nno" (default) for optimized NNM or "quickmatch".}
}
\value{
A quickmatch object containing:

\code{match.matrix}:   a \code{(controls+1)} columns matrix, each row containing the numbers
of a matched treated (in col 1) and corresponding
control (in columnss \code{2:(controls+1))} objects. The numbers
of objects are positions in vector x.
If, for a treated object, there are only \code{0<k<controls}
matches, the elements of the row in the
cols (k+2):(controls+1) are filled with integer NAs.

\code{num_pairs}:   vector of length total.matches showing the number of
controls matched to each treated object. \code{num_pairs[j]}
corresponds to the j-th row of match.matrix.

\code{discarded.t}:   vector of the numbers (positions in x)
of treated objects which are not matched.

\code{discarded.c}:   vector of the numbers (positions in x)
of control objects which are not matched.

\code{total.matches}:  the number of matched treated objects, which is the
number of rows in match.matrix.

\code{total.pairs}:   the number of matched control objects, which is the
number of non-NA elements in columns \code{2:(controls+1)}
of match.matrix
}
\description{
Given treatment and control groups, finds a pairing of treatment units
to controls that maximizes number of matched disjoint pairs satisfying a
given caliper.
}
\details{
Quickmatch is generic function, calls one of its methods.
}
\examples{
quickmatch(c(1,1.1,2.5,1.5,0.2,2,0.5,2),c(0,1,0,1,0,1,0,1),0.5)
}
\references{
P.S. Ruzankin (2019). A fast algorithm for maximal propensity score matching.
\emph{Methodol Comput Appl Probab}. \url{https://doi.org/10.1007/s11009-019-09718-4}
}
\author{
Pavel S. Ruzankin,
}
