% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{treatment_effects}
\alias{treatment_effects}
\title{Average treatment effect estimator in matched groups}
\usage{
treatment_effects(outcomes, treatments, matching, contrasts = NULL,
  subset = NULL, drop = TRUE)
}
\arguments{
\item{outcomes}{numeric vector with observed outcomes.}

\item{treatments}{integer or factor vector with treatment indicators.}

\item{matching}{qm_matching object containing the matching.}

\item{contrasts}{vector of treatment labels (corresponding to \code{treatments})
specifying which treatment effects to estimate. If \code{NULL},
all treatment effects will be estimated.}

\item{subset}{units to estimate potential outcomes for. If \code{NULL},
the estimate will pertain to all units in the sample (i.e.,
corresponding to ATE). A non-null value specificies a subset
of units that the estimate should pertain to (e.g., ATT or ATC).
If \code{subset} is a logical vector and of length equal to the
sample size, units indicated with \code{TRUE} will be included.
Otherwise, \code{subset} should contain treatment labels and
the corresponding units (as given by \code{treatments}) will be
included.}

\item{drop}{if \code{FALSE}, the function always returns a matrix (even when
\code{contrasts} contains two elements).}
}
\value{
Returns the estimated treatment effects. If \code{contrasts} is of length
        two, a scaler is returned with the corresponding treatment effect. In all
        other cases, a numeric matrix with all treatment effects are returned. Rows
        in this matrix indicate minuends in the treatment contrast and columns
        indicate subtrahends.
}
\description{
\code{treatment_effects} estimates treatment effect in matched groups. For a
set of treatment conditions, the function returns estimates of treatment
effects for the units in the sample for each pair of conditions.
It is possible to estimate treatment effects for a subset of the
units (e.g., ATT or ATC).
}

