% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm_matching.R
\name{is.qm_matching}
\alias{is.qm_matching}
\title{Check \code{qm_matching} object}
\usage{
is.qm_matching(obj)
}
\arguments{
\item{obj}{object to check.}
}
\value{
Returns \code{TRUE} if \code{obj} is a valid
        \code{qm_matching} object, otherwise \code{FALSE}.
}
\description{
\code{check_qm_matching} checks whether the provided object
is a valid instance of the \code{qm_matching} class. It does
not whether the matching is sensible.
}

