% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm_matching.R
\name{is.qm_matching}
\alias{is.qm_matching}
\title{Check qm_matching object}
\usage{
is.qm_matching(x)
}
\arguments{
\item{x}{object to check.}
}
\value{
Returns \code{TRUE} if \code{x} is a valid \code{\link{qm_matching}}
   object, otherwise \code{FALSE}.
}
\description{
\code{is.qm_matching} checks whether the provided object is a valid instance
of the \code{\link{qm_matching}} class.
}
\details{
\code{is.qm_matching} does not check whether the matching itself is sensible
or whether it satisfies some set of constraints. See
\code{\link[scclust]{check_clustering}} for that functionality.
}
