% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{potential_outcomes}
\alias{potential_outcomes}
\title{Average potential outcome estimator in matched groups}
\usage{
potential_outcomes(outcomes, treatments, matching, estimands = NULL,
  subset = NULL)
}
\arguments{
\item{outcomes}{numeric vector with observed outcomes.}

\item{treatments}{integer or factor vector with treatment indicators.}

\item{matching}{qm_matching object containing the matching.}

\item{estimands}{vector of treatment labels (corresponding to \code{treatments})
specifying which potential outcomes to estimate. If \code{NULL},
all potential outcomes will be estimated.}

\item{subset}{units to estimate potential outcomes for. If \code{NULL},
the estimate will pertain to all units in the sample (i.e.,
corresponding to ATE). A non-null value specificies a subset
of units that the estimate should pertain to (e.g., ATT or ATC).
If \code{subset} is a logical vector and of length equal to the
sample size, units indicated with \code{TRUE} will be included.
Otherwise, \code{subset} should contain treatment labels and
the corresponding units (as given by \code{treatments}) will be
included.}
}
\value{
Returns a named numeric vector with an estimate of each potential outcome.
}
\description{
\code{potential_outcomes} estimates potential outcomes using matched groups. For
an outcome, a matching and treatments, the function returns estimates of the
average potential outcomes for all treatments for the units in the sample. It is
possible to estimate potential outcomes for a subset of the units (e.g., to derive
ATT or ATC).
}

