% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm_matching.R
\name{qm_matching}
\alias{qm_matching}
\title{Construct a \code{qm_matching} object}
\usage{
qm_matching(labels, unassigned_labels = NULL, ids = NULL)
}
\arguments{
\item{labels}{a vector containing each data point's label.}

\item{unassigned_labels}{a vector containing labels that denote unassigned data points.}

\item{ids}{optional IDs for the data points. Either a vector of the same length as
\code{cluster_labels} or \code{NULL}. If \code{NULL}, the IDs are set to
\code{1:length(labels)}.}
}
\value{
Returns a \code{qm_matching} object.
}
\description{
\code{qm_matching} constructs a \code{qm_matching} object from existing
matching labels.
}
\details{
\code{qm_matching} inherits from \code{Rscc_clustering}.
}

