% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickmatch-package.R
\docType{package}
\name{quickmatch-package}
\alias{quickmatch-package}
\title{quickmatch: Quick Generalized Full Matching}
\description{
Provides functions for constructing near-optimal generalized full matchings.
Generalized full matching is an extension of the original full matching method
to situations with more intricate study designs. The package is made with
large data sets in mind and derives matchings more than an order of magnitude
quicker than other methods.
}
\details{
See \code{\link{quickmatch}} for the main matching function.

See the package's website for more information:
\url{https://github.com/fsavje/quickmatch}.

Bug reports and suggestions are greatly appreciated. They are best reported
here: \url{https://github.com/fsavje/quickmatch/issues}.
}
\references{
Sävje, Fredrik and Michael J. Higgins and Jasjeet S. Sekhon (2017),
\sQuote{Generalized Full Matching}, arXiv 1703.03882.
\url{https://arxiv.org/abs/1703.03882}
}
