% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickmatch.R
\name{quickmatch}
\alias{quickmatch}
\title{Construct matchings}
\usage{
quickmatch(distances, treatments, treatment_constraints = NULL,
  total_size_constraint = NULL, caliper = NULL, subset = NULL, ...)
}
\arguments{
\item{distances}{abc}

\item{treatments}{abc}

\item{treatment_constraints}{abc}

\item{total_size_constraint}{abc}

\item{caliper}{abc}

\item{subset}{abc}

\item{...}{abc}
}
\value{
abc
}
\description{
abc
}

