% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickmatch.R
\name{quickmatch}
\alias{quickmatch}
\title{Construct matched groups}
\usage{
quickmatch(distances, treatments, treatment_constraints = NULL,
  total_size_constraint = NULL, subset = NULL, caliper = NULL, ...)
}
\arguments{
\item{distances}{an \code{Rscc_distances} object containing distances between
the units in the sample (i.e., measures on how similar the
units are). See \code{\link[Rscclust]{make_distances}} for
details on how to construct this object.}

\item{treatments}{integer or factor vector with treatment indicators.}

\item{treatment_constraints}{a named integer vector containing the matching constraints. If
\code{NULL}, the function defaults to requiring at least one unit
of each treatment condition in each matched group.}

\item{total_size_constraint}{an integer with the total size constraint of the matched
groups.}

\item{subset}{units to target the matching for. All units indicated by \code{subset}
are ensured to be assigned to a matched group (disregarding eventual
\code{caliper} setting). Units not indicated by \code{subset} could
be left unassigned if they are not necessary to satisfy the matching
constraints. If \code{NULL}, \code{quickmatch} targets all units and
ensures that all units are assigned to a group. If \code{subset} is
a logical vector and of length equal to the
sample size, units indicated with \code{TRUE} will be included.
Otherwise, \code{subset} should contain treatment labels, and
the corresponding units (as given by \code{treatments}) will be
included.}

\item{caliper}{impose a caliper on the matching that restricts the maximum
allowed distance between units in the matching. The caliper
is imposed in the underlying data structure used by the algorithm,
and it does not directly correspond the maximum within-group
distance. See below for details.}

\item{...}{additional parameters to be sent the underlying
\code{\link[Rscclust]{nng_clustering_types}} function.}
}
\value{
Returns a \code{\link{qm_matching}} object describing the matched groups.
}
\description{
\code{quickmatch} constructs matched groups that satisfy matching constraints.
}
\examples{
my_data <- data.frame(y = rnorm(100),
                      x1 = runif(100),
                      x2 = runif(100),
                      treatment = factor(sample(rep(c("T", "C"), 50))))

}
\seealso{
See \code{\link[Rscclust]{nng_clustering_types}} for the underlying function used to
         construct the matched groups. See \code{\link{potential_outcomes}} and
         \code{\link{treatment_effects}} for estimators that can be used with the produced matching.
}

