% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_estimator.R
\name{regression_estimator}
\alias{regression_estimator}
\title{Regression-based average treatment effect estimator}
\usage{
regression_estimator(outcomes, treatments, matching, covariates = NULL,
  subset = NULL)
}
\arguments{
\item{outcomes}{numeric vector with observed outcomes.}

\item{treatments}{factor specifying which treatments the units are assigned to.}

\item{matching}{\code{\link{qm_matching}} or \code{\link[scclust]{scclust}} object with
the matched groups.}

\item{covariates}{vector, matrix or data frame with covariates to include in the estimation.
If \code{NULL}, no covariates are included.}

\item{subset}{units to target the estimation for. If \code{NULL}, the effect is estimated
for all units in the sample (i.e., ATE). A non-null value
specifies a subset of units that the effect should be estimated for (e.g.,
ATT or ATC). If \code{subset} is a logical vector with the same length as
the sample size, units indicated with \code{TRUE} will be targeted. If
\code{subset} is an integer vector, the units with indices in \code{subset}
are targeted. If
\code{subset} is a character vector, it should contain treatment labels,
and the effect for the corresponding units (as given by \code{treatments}) will be
estimated.}
}
\value{
A list with two numeric matrices with all estimated treatment effects
   and their estimated variances is returned.
   The first matrix (\code{effects}) contains estimated treatment effects.
   Rows in this matrix indicate minuends in the treatment effect contrast and
   columns indicate subtrahends. For example, in the matrix:
   \tabular{rrrr}{
     \tab a \tab b \tab c\cr
     a \tab 0.0 \tab 4.5 \tab 5.5\cr
     b \tab -4.5 \tab 0.0 \tab 1.0\cr
     c \tab -5.5 \tab -1.0 \tab 0.0\cr
   }
   the estimated treatment effect between conditions \eqn{a} and \eqn{b} is
   \eqn{4.5}, and the estimated treatment effect between conditions \eqn{c}
   and \eqn{b} is \eqn{-1.0}. Or in symbols \eqn{E[Y(a) - Y(b) | S] = 4.5} and
   \eqn{E[Y(c) - Y(b) | S] = -1.0} where \eqn{S} is the condition set indicated
   by the \code{subset} parameter.

   The second matrix (\code{effect_variances}) contains estimates of
   variances of the corresponding effect estimators.
}
\description{
\code{regression_estimator} estimates treatment effects in matched samples.
The function expects the user to provide the outcomes, treatment indicators, and
a matching object. The function returns point estimates of the average treatment
effects and variance estimates. It is possible to estimate treatment effects for
subsets of the observations, such as estimates of the average treatment effect for
the treated.
}
\details{
\code{regression_estimator} estimates treatment effects using weighted
regression. The function first derives unit-level weights implied by the
matching. In detail, let \eqn{S(g)} be the number of unit indicated by
\code{subset} in group \eqn{g} (or the total number of units in the group
if \code{subset} is \code{NULL}). Let \eqn{T} be the total number of units
indicated by \code{subset} in sample (or the sample size if \code{subset}
is \code{NULL}). Let \eqn{A(t, g)} be the number of units assigned to treatment \eqn{t} in group \eqn{g}. The weight for a unit in group
\eqn{g} that is assigned to treatment \eqn{t} is given by:

\deqn{\frac{S(g)}{T \times A(t, g)}.}{S(g) / [T * A(t, g)].}

For example, consider a matched group with one treated unit and two control
units when we are interested in the average effect of the treated (ATT) and
we have 50 treated units in total (\eqn{T=50}). For all three units in the
group, we have \eqn{S(g)=1}. For the treated unit we have \eqn{A(t, g)=1},
so its weight becomes \eqn{1/50}. The two control units have \eqn{A(t, g)=2},
so their weights are both \eqn{1/100}.

These weights are such that the difference between the weighted outcome
averages of two treatment conditions is the same as the average within-group
difference-in-means between the two conditions.

The function uses the derived weights in a weighted least squares regression
(using the \code{\link[stats]{lm}} function) with indicator variables for the treatment
conditions. Optionally, covariates can be added to the regression as well (e.g., a common
recommendation is to include the covariates used to construct the matching).
Standard errors are estimated with the heteroskedasticity-robust estimates derived
using the "HC1" estimator in the \code{\link[sandwich]{vcovHC}} function.
Units assigned a weight of zero is excluded from the estimation.
}
\examples{
# Construct example data
my_data <- data.frame(y = rnorm(100),
                      x1 = runif(100),
                      x2 = runif(100),
                      treatment = factor(sample(rep(c("T1", "T2", "C"), c(25, 25, 50)))))

# Make distances
my_distances <- distances(my_data, dist_variables = c("x1", "x2"))

# Make matching
my_matching <- quickmatch(my_distances, my_data$treatment)

# ATE without covariates
regression_estimator(my_data$y,
                     my_data$treatment,
                     my_matching)

# ATE with covariates
regression_estimator(my_data$y,
                     my_data$treatment,
                     my_matching,
                     my_data[c("x1", "x2")])

# ATT for T1
regression_estimator(my_data$y,
                     my_data$treatment,
                     my_matching,
                     my_data[c("x1", "x2")],
                     subset = "T1")

}
\references{
Stuart, Elizabeth A. (2010),
   \sQuote{Matching Methods for Causal Inference: A Review and a Look Forward}.
   Statistical Science, 25(1), 1–21. https://doi.org/10.1214/09-STS313
}
